Move bookings component into customer module and set up click handlers
