+++ thesis-app/src/app/customer/components/booking/booking.component.html
@@ -1,0 +1,3 @@
+<h4>{{name}}</h4>
+<span class="glyphicon glyphicon-time">{{time}}</span>
+<span class="glyphicon glyphicon-home">{{phonenumber}}</span>
