+++ thesis-app/src/app/customer/components/landing/landing.component.html
@@ -1,42 +1,93 @@
 <customer-navbar></customer-navbar>
 <h1>Appointments</h1>
 
-<div *ngIf="profile.type===0">
-  <button [routerLink]="['/stylisthome']">Find Stylist</button>
-</div>
 <div *ngIf="profile.type===1">
   <button [routerLink]="['/home']">Find Stylist</button>
-</div>
 
-<h2>Confirmed</h2>
-<div
-  *ngFor="let confirmedBooking of profile.confirmedBookings; let i = index"
-  class="booking-confirmed"
-  (click)="removeConfirmedBooking(confirmedBooking.id, i)"
->
-  {{confirmedBooking.name}}
-  {{confirmedBooking.time}}
-  {{confirmedBooking.phonenumber}}
-</div>
-
-<h2>Awaiting Payment</h2>
-<div *ngFor="let dueBooking of profile.dueBookings">
-  {{dueBooking.name}}
-  {{dueBooking.time}}
-  {{dueBooking.phonenumber}}
-  <payment-input
-    (handleProcessedPayment)="completeBooking($event)"
-    [stylistName]="dueBooking.name"
+  <h2>Confirmed</h2>
+  <div
+    *ngFor="let confirmedBooking of profile.confirmedBookings; let i = index"
+    class="booking-confirmed"
+    (click)="removeConfirmedBooking(confirmedBooking.id, i)"
   >
-  </payment-input>
+    <booking
+      [name]="confirmedBooking.name"
+      [time]="confirmedBooking.time"
+      [phoneNumber]="confirmedBooking.phonenumber"
+    >
+    </booking>
+  </div>
+
+  <h2>Awaiting Payment</h2>
+  <div *ngFor="let dueBooking of profile.dueBookings">
+    <booking
+      [name]="dueBooking.name"
+      [time]="dueBooking.time"
+      [phoneNumber]="dueBooking.phonenumber"
+    >
+    </booking>
+    <payment-input
+      (handleProcessedPayment)="completeBooking($event)"
+      [stylistName]="dueBooking.name"
+    >
+    </payment-input>
+  </div>
+
+  <h2>Pending</h2>
+  <div
+    *ngFor="let pendingBooking of profile.pendingBookings; let i = index"
+    class="booking-confirmed"
+  >
+    <booking
+      [name]="pendingBooking.name"
+      [time]="pendingBooking.time"
+      [phoneNumber]="pendingBooking.phonenumber"
+    >
+    </booking>
+    <button (click)="deleteBooking(pendingBooking.id, i)">Reject</button>
+  </div>
 </div>
 
-<h2>Pending</h2>
-<div
-  *ngFor="let pendingBooking of profile.pendingBookings; let i = index"
-  class="booking-confirmed"
->
-  {{pendingBooking.name}}
-  {{pendingBooking.time}}
-  {{pendingBooking.phonenumber}}
+<div *ngIf="profile.type===0">
+  <button [routerLink]="['/stylisthome']">Find Stylist</button>
+
+  <h2>Confirmed</h2>
+  <div
+    *ngFor="let confirmedBooking of profile.confirmedBookings; let i = index"
+    class="booking-confirmed"
+    (click)="removeConfirmedBooking(confirmedBooking.id, i)"
+  >
+    <booking
+      [name]="confirmedBooking.name"
+      [time]="confirmedBooking.time"
+      [phoneNumber]="confirmedBooking.phonenumber"
+    >
+    </booking>
+  </div>
+
+  <h2>Awaiting Payment</h2>
+  <div *ngFor="let dueBooking of profile.dueBookings">
+    <booking
+      [name]="dueBooking.name"
+      [time]="dueBooking.time"
+      [phoneNumber]="dueBooking.phonenumber"
+    >
+    </booking>
+    <button>Request Payment</button>
+  </div>
+
+  <h2>Pending</h2>
+  <div
+    *ngFor="let pendingBooking of profile.pendingBookings; let i = index"
+    class="booking-confirmed"
+  >
+    <booking
+      [name]="pendingBooking.name"
+      [time]="pendingBooking.time"
+      [phoneNumber]="pendingBooking.phonenumber"
+    >
+    </booking>
+    <button (click)="confirmBooking(pendingBooking.id, i)">Confirm</button>
+    <button (click)="deleteBooking(pendingBooking.id, i)">Reject</button>
+  </div>
 </div>
