--- thesis-app/src/app/stylist/components/booking/booking.component.html
@@ -1,13 +1,0 @@
-<div class="is-confirmed" *ngIf="bookingInformation.isconfirmed===1">
-  <h4>{{bookingInformation.customer}}</h4>
-  <span class="glyphicon glyphicon-time">{{bookingInformation.time}}</span>
-  <span class="glyphicon glyphicon-home">{{bookingInformation.location}}</span>
-  <span class="complete-btn" (click)="completeBooking.emit(bookingInformation.id)">Get Paid</span>
-</div>
-<div class="not-confirmed" *ngIf="bookingInformation.isconfirmed===0">
-  <h4>{{bookingInformation.customer}}</h4>
-  <span class="glyphicon glyphicon-time">{{bookingInformation.time}}</span>
-  <span class="glyphicon glyphicon-home">{{bookingInformation.location}}</span>
-  <span class="confirm-btn" (click)="confirmBooking.emit(bookingInformation.id)">Confirm</span>
-  <span class="delete-btn" (click)="deleteBooking.emit(bookingInformation.id)">Delete</span>
-</div>
