+++ thesis-app/src/app/stylist/components/index.ts
@@ -1,3 +1,2 @@
 export * from './stylist-home';
 export * from './stylist-map';
-export * from './booking';
