+++ thesis-app/src/app/stylist/components/stylist-home/stylist-home.component.html
@@ -1,13 +1,5 @@
 <customer-navbar></customer-navbar>
 <div class="wrapper">
-  <div class="booking-container" *ngFor="let booking of bookings; let i = index;">
-    <booking
-      [bookingInformation]="booking"
-      (confirmBooking)="confirmBooking($event, i)"
-      (deleteBooking)="deleteBooking($event, i)"
-      (completeBooking)="completeBooking($event, i)"
-    ></booking>
-  </div>
   <div class="map-container">
     <stylist-map class="map"
       [stylistId]="stylistProfile.id"
